Adds filters to shortcircuit the GS public API
By doing this, we allow both the Gutenberg plugin and any 3rd party
to change the behavior of these functions. This is useful specially
for Gutenberg, where we may need to introduce future changes so
we need a way to hook into the core functions.

It's inspired by the pre_render_block filter.